requirements.txt
pandas
psycopg2-binary
boto3
simple-salesforce
